extractFromResource is a parser and should be with other parsers
